# backend/src/rest/v1/openapi.yaml
openapi: 3.0.3
info:
  title: ChatFlow API
  description: A minimal but clean WhatsApp-style messaging app API
  version: 1.0.0
  contact:
    name: ChatFlow Team
servers:
  - url: http://localhost:3002/v1
    description: Development server
  - url: https://chatflow-backend-3w6u4kmniq-ue.a.run.app/v1
    description: Production server (Cloud Run)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Conversation:
      type: object
      properties:
        id:
          type: string
          pattern: '^conv_[0-9]+_[a-z0-9]+$'
          description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
          example: 'conv_1750386041311_fpmswok2p'
        createdBy:
          type: string
          format: email
        type:
          type: string
          enum: [DIRECT, GROUP]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ConversationParticipant'
...
    ConversationParticipant:
      type: object
      properties:
        userId:
          type: string
          format: email
        joinedAt:
          type: string
          format: date-time
        role:
          type: string
          enum: [ADMIN, MEMBER]
    PaginationResult:
      type: object
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data: {}
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details: {}
...
    CreateConversationRequest:
      type: object
      required:
        - participantEmails
      properties:
        participantEmails:
          type: array
          items:
            type: string
            format: email
          minItems: 1
...
    SearchResult:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
          format: email
        content:
          type: string
        messageType:
          type: string
          enum: [TEXT, IMAGE, FILE]
        createdAt:
          type: string
          format: date-time
        relevanceScore:
          type: number
          minimum: 0
          maximum: 1
        highlights:
          type: array
          items:
            type: string
        conversationContext:
          type: object
          properties:
            participants:
              type: array
              items:
                type: string
            createdAt:
              type: string
              format: date-time
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        query:
          type: string
        totalResults:
          type: integer
        searchTime:
          type: integer
          description: Search execution time in milliseconds
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details: {}
...
    KeyMetadata:
      type: object
      properties:
        keyId:
          type: string
          description: Unique key identifier
        userId:
          type: string
          format: email
          nullable: true
          description: User associated with the key (null for system keys)
        purpose:
          type: string
          enum: [message, search, suggestion, general]
          description: Purpose of the key
        algorithm:
          type: string
          enum: [AES-256-GCM]
          description: Encryption algorithm
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        rotatedFrom:
          type: string
          nullable: true
          description: Previous key ID if this is a rotation
        isActive:
          type: boolean
        version:
          type: integer
          minimum: 1
...
    CurrentKeyIds:
      type: object
      properties:
        keyIds:
          type: object
          properties:
            message:
              type: string
            search:
              type: string
            suggestion:
              type: string
        version:
          type: string
        lastUpdated:
          type: string
          format: date-time
        instructions:
          type: object
          properties:
            usage:
              type: string
            derivation:
              type: string
            security:
              type: string
...
    EncryptionMetadata:
      type: object
      description: "Metadata for field-level encryption using AES-256-GCM"
      required:
        - algorithm
        - keyId
        - iv
        - tag
      properties:
        algorithm:
          type: string
          enum: ["AES-256-GCM"]
          description: "Encryption algorithm used"
          example: "AES-256-GCM"
        keyId:
          type: string
          description: "Client-side key identifier for key management"
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "user_key_001"
        iv:
          type: string
          format: base64
          description: "Base64 encoded initialization vector (96-bit for GCM)"
          example: "MTIzNDU2Nzg5MGFiY2RlZg=="
        tag:
          type: string
          format: base64
          description: "Base64 encoded authentication tag (128-bit for GCM)"
          example: "dGFnMTIzNDU2Nzg5MGFiY2Q="
        timestamp:
          type: integer
          format: int64
          description: "Unix timestamp for replay attack prevention"
          example: 1640995200
        nonce:
          type: string
          description: "Unique nonce for each encryption operation"
          example: "unique_nonce_001"
    EncryptedField:
      type: object
      description: "Encrypted field with metadata"
      required:
        - data
        - encryption
      properties:
        data:
          type: string
          format: base64
          description: "Base64 encoded encrypted data"
          example: "eyJhbGciOiJBMjU2R0NNIiwiZW5jIjoiQTI1NkdDTSJ9..."
        encryption:
          $ref: '#/components/schemas/EncryptionMetadata'
paths:
...
  /conversations:
    get:
      tags:
        - Conversations
      summary: Get user's conversations
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Conversations retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PaginationResult'
                          - type: object
                            properties:
                              data:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Conversation'

    post:
      tags:
        - Conversations
      summary: Create a new conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Conversation'
...
  /search/conversations:
    get:
      tags:
        - Search
      summary: Search conversations with semantic understanding
      description: Perform intelligent search across conversations using natural language queries
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 500
          description: Search query (supports natural language)
          example: "lunch plans with Sarah"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          description: Maximum number of results to return
      responses:
        '200':
          description: Search results with semantic relevance scoring
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Search service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Search
      summary: Search conversations (POST version for web clients)
      description: Same functionality as GET but via POST to potentially reduce OPTIONS preflight overhead in some scenarios
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - q
              properties:
                q:
                  oneOf:
                    - type: string
                      minLength: 1
                      maxLength: 500
                      description: "Plain text search query (not recommended for production)"
                      example: "lunch plans with Sarah"
                    - $ref: '#/components/schemas/EncryptedField'
                  description: "Search query - should be encrypted for privacy"
                limit:
                  type: integer
                  minimum: 1
                  maximum: 50
                  default: 20
                  description: Maximum number of results to return
      responses:
        '200':
          description: Search results with semantic relevance scoring
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Search service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /keys/current:
    get:
      tags:
        - Key Coordination
      summary: Get current keyIds for encryption (no actual keys)
      description: Retrieve current keyIds that clients should use for encryption without exposing actual keys
      responses:
        '200':
          description: Current keyIds for client-side encryption
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CurrentKeyIds'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/keys:
    get:
      tags:
        - Admin - Key Management
      summary: List key metadata (no actual keys)
      description: List key metadata with filtering options (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: purpose
          in: query
          schema:
            type: string
            enum: [message, search, suggestion, general]
          description: Filter by key purpose
        - name: userId
          in: query
          schema:
            type: string
            format: email
          description: Filter by user ID
        - name: isActive
          in: query
          schema:
            type: boolean
          description: Filter by active status
        - name: includeExpired
          in: query
          schema:
            type: boolean
            default: false
          description: Include expired keys in results
      responses:
        '200':
          description: List of key metadata
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/KeyMetadata'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...